Fix getting older items with non-successive IDs
Bug
---

set = [ 1, 2, 8 ]
batch_size = 5

A query with client#get for older items from id '8', the reply
would be [ 8 ], as the client internally looks at the range 3 - 8,
that is, the range is calculated by subtracting the batch size from
the id '8' (max) i.e. 8 - 5 = 3

The expected reply would be [ 1, 2, 8 ] so that the missing ids are
ignored. Regardless of the succession of ids, client#get should
return the previous 'n' or next 'n' of items.